requirements.txt
fastapi==0.114.1
sqlalchemy==2.0.34
pymysql==1.1.1
mysql.connector.python==9.0.0
python-decouple==3.8
SQLModel==0.0.22
cryptography==43.0.1
uvicorn==0.30.6